pyproject.toml
[project]
name = "arnold"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "wandb>=0.22.3,<0.23",
    "sk-video==1.1.10",
    "gymnasium==0.29.1",
    "gym==0.26.2",
    "gym-notices==0.0.8",
    "hydra-core>=1.3.2,<2.0",
    "mujoco==3.3.6",
    "easydict==1.11",
    "h5py==3.10.0",
    "importlib-resources==6.4.5",
    "ipdb==0.13.13",
    "joblib>=1.5.2,<2.0",
    "matplotlib>=3.10.7,<4.0",
    "numpy>=2.3.4,<3.0",
    "numpy-stl>=3.2.0,<4.0",
    "omegaconf==2.3.0",
    "pandas>=2.3.3,<3.0",
    "pyopengl==3.1.0",
    "pytest>=8.4.2,<9.0",
    "scikit-learn>=1.7.2,<2.0",
    "scipy>=1.16.3,<2.0",
    "termcolor>=3.2.0,<4.0",
    "torch>=2.9.0,<3.0",
    "tqdm>=4.67.1,<5.0",
    "ipykernel>=7.1.0,<8.0",
    "smplx @ git+https://github.com/ZhengyiLuo/smplx.git@a5b8e4ac14f79f3f33fd2cf2a16e6f507146b813",
    "lxml>=6.0.2,<7.0",
    "httpx>=0.28.1,<0.29",
    "openai>=2.8.1,<3.0",
    "huggingface-hub>=1.2.3,<2.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
